+++ lisp/eab-compile.el
@@ -93,10 +93,10 @@
       (save-window-excursion
 	(eab/compile eab/gr-command))
       (setq eab/gr-ready? 't))
-    (eab/async-update-gr)
+    (eab/async-update-gr :recompile 't :notify 't)
     (switch-to-buffer eab/gr-buffer nil 't)))
 
-(cl-defun eab/async-update-gr (&optional &key recompile)
+(cl-defun eab/async-update-gr (&optional &key recompile notify)
   (async-start
    `(lambda ()
       (with-temp-buffer
@@ -108,15 +108,30 @@
 	    (switch-to-buffer eab/gr-buffer)
 	    (let ((compilation-buffer-name-function nil))
 	      (recompile))))
+      '(if ,notify
+	  (if (> (string-to-number (shell-command-to-string eab/check-gr-command)) 6)
+	      (eab/gotify "fz" "updated" 0)))
       (message "async result: <%s>" result))))
 
 (defun eab/update-gr-status-on-idle ()
   (eab/async-update-gr :recompile 't))
 
 (setq eab/gotify-ready? nil)
+(defvar eab/gotify-websocket nil "gotify websocket")
 (defun eab/gotify-status (&optional arg)
   (interactive)
   (let ((gotify-buffer "*compilation: gotify*"))
+    (unless (websocket-openp eab/gotify-websocket)
+      (setq eab/gotify-websocket
+            (websocket-open
+	     eab/gotify-ws
+             :on-message `(lambda (_websocket frame)
+			    (let ((text (websocket-frame-text frame)))
+			      (save-window-excursion
+				(switch-to-buffer ,gotify-buffer)
+				(let ((compilation-buffer-name-function nil))
+				  (recompile)))))
+             :on-close (lambda (_websocket) (message "websocket closed")))))
     (if eab/gotify-ready?
 	(switch-to-buffer gotify-buffer nil 't)
       (progn
@@ -124,20 +139,13 @@
 	       `(lambda (mode) ,gotify-buffer)))
 	  (save-window-excursion
 	    (eab/compile eab/gotify-command)))
-	(setq eab/websocket
-              (websocket-open eab/gotify-ws
-                              :on-message `(lambda (_websocket frame)
-					    (save-window-excursion
-					      (switch-to-buffer ,gotify-buffer)
-					      (let ((compilation-buffer-name-function nil))
-						(recompile))))
-                              ;; (message "ws frame: %S" (websocket-frame-text frame)))
-                              :on-close (lambda (_websocket) (message "websocket closed"))))
 	(setq eab/gotify-ready? 't)
 	(switch-to-buffer gotify-buffer nil 't)))))
 
-;; (websocket-send-text eab/websocket "hello from emacs")
-;; (websocket-close eab/websocket)
+'((websocket-send-text eab/gotify-websocket "hello from emacs")
+  (websocket-close eab/gotify-websocket)
+  (websocket-openp eab/gotify-websocket)
+  )
 
 (defun eab/compile-goto-error ()
   (interactive)